[CDSR-3017] add EORI = TPI02500 test case for simulating EIS 500 error

+++ app/uk/gov/hmrc/cdsreimbursementclaimstubs/models/ReasonForSecurity.scala
@@ -42,10 +42,10 @@ object ReasonForSecurity {
       "ACS" // placeholder for future RfS code additions
     )
 
-  def ofIndex(index: Int): String =
-    values(Math.abs(index) % values.size)
+  def ofIndex(index: Long): String =
+    values((Math.abs(index) % values.size).toInt)
 
-  def of(value: String): String   =
+  def of(value: String): String    =
     ofIndex(value.filter(Character.isDigit).toInt)
 
 }
